+++ meos/src/arrow/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(ARROW_SOURCES
+  temporal_arrow.c
+)
+
+add_library(arrow OBJECT ${ARROW_SOURCES})
