+++ meos/src/temporal/type_in.c
@@ -41,8 +41,8 @@
 #include "utils/timestamp.h"
 /* MEOS */
 #include <meos.h>
-#include <meos_rgeo.h>
 #include <meos_internal.h>
+#include <meos_internal_geo.h>
 #include "temporal/set.h"
 #include "temporal/span.h"
 #include "temporal/tbox.h"
@@ -65,6 +65,7 @@
   #include "pose/pose.h"
 #endif
 #if RGEO
+  #include <meos_rgeo.h>
   #include "rgeo/trgeo.h"
 #endif
