+++ mobilitydb/sql/geo/053_tpoint_inout.in.sql
@@ -214,6 +214,7 @@ CREATE FUNCTION asHexEWKB(tgeogpoint, endianenconding text DEFAULT '')
   AS 'MODULE_PATHNAME', 'Tspatial_as_hexewkb'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
+#if ARROW
 CREATE FUNCTION arrowRoundtrip(tgeompoint)
   RETURNS tgeompoint
   AS 'MODULE_PATHNAME', 'Temporal_arrow_roundtrip'
@@ -222,5 +223,6 @@ CREATE FUNCTION arrowRoundtrip(tgeogpoint)
   RETURNS tgeogpoint
   AS 'MODULE_PATHNAME', 'Temporal_arrow_roundtrip'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+#endif //ARROW
 
 /*****************************************************************************/
