+++ mobilitydb/sql/rgeo/150_trgeo.in.sql
@@ -707,7 +707,9 @@ CREATE OPERATOR CLASS trgeometry_hash_ops
 
 /******************************************************************************/
 
+#if ARROW
 CREATE FUNCTION arrowRoundtrip(trgeometry)
   RETURNS trgeometry
   AS 'MODULE_PATHNAME', 'Temporal_arrow_roundtrip'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+#endif //ARROW
