+++ mobilitydb/test/rgeo/CMakeLists.txt
@@ -26,6 +26,12 @@ set(SKIP_TESTS
 foreach(file ${testfiles})
   get_filename_component(TESTNAME ${file} NAME_WE)
   set(DOTEST TRUE)
+  if(${TESTNAME} MATCHES "_arrow")
+    if(NOT ARROW)
+      message("Disabling test ${TESTNAME} (ARROW=OFF)")
+      set(DOTEST FALSE)
+    endif()
+  endif()
   if(${TESTNAME} MATCHES "_pg([0-9]+)")
     if(${POSTGRESQL_VERSION_MAJOR} LESS ${CMAKE_MATCH_1})
       message("Disabling test ${TESTNAME}")
